Finished lana main, added boss mechanics, enabled mega

+++ include/pokemon.h
@@ -362,7 +362,7 @@ u8 GetPlayerFlankId(void);
 u16 GetLinkTrainerFlankId(u8 id);
 s32 GetBattlerMultiplayerId(u16 a1);
 u8 GetTrainerEncounterMusicId(u16 trainerOpponentId);
-u16 ModifyStatByNature(u8 nature, u16 n, u8 statIndex);
+u32 ModifyStatByNature(u8 nature, u32 n, u8 statIndex);
 void AdjustFriendship(struct Pokemon *mon, u8 event);
 void MonGainEVs(struct Pokemon *mon, u16 defeatedSpecies);
 u16 GetMonEVCount(struct Pokemon *mon);
