+++ src/script_pokemon_util_80F87D8.c
@@ -649,7 +649,7 @@ void CreateScriptedWildMon(u16 species, u8 level, u16 item)
     }
 }
 
-void ScriptSetMonMoveSlot(u8 monIndex, u16 move, u8 slot)
+void ScriptSetMonMoveSlot(u16 monIndex, u16 move, u16 slot)
 {
     if (monIndex > PARTY_SIZE)
         monIndex = gPlayerPartyCount - 1;
